minor adjustments and cleanup
