+++ src/login/mobile-token-onboarding/MobileToken.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import {View} from 'react-native';
-import {TCardAsMobileToken} from '@atb/login/travel-document-onboarding/TCardAsMobileToken';
-import {PhoneAsMobileToken} from '@atb/login/travel-document-onboarding/PhoneAsMobileToken';
+import {TCardAsMobileToken} from '@atb/login/mobile-token-onboarding/TCardAsMobileToken';
+import {PhoneAsMobileToken} from '@atb/login/mobile-token-onboarding/PhoneAsMobileToken';
 import {StyleSheet} from '@atb/theme';
 import {StaticColorByType} from '@atb/theme/colors';
 import {useMobileTokenContextState} from '@atb/mobile-token/MobileTokenContext';
-import {NoMobileToken} from '@atb/login/travel-document-onboarding/NoMobileToken';
+import {NoMobileToken} from '@atb/login/mobile-token-onboarding/NoMobileToken';
 
 const themeColor: StaticColorByType<'background'> = 'background_accent_0';
 
@@ -13,7 +13,7 @@ const MobileToken = () => {
   const styles = useStyles();
   const {travelTokens} = useMobileTokenContextState();
   const inspectableToken = travelTokens?.find((t) => t.inspectable)!;
-  switch (inspectableToken.type) {
+  switch (inspectableToken?.type) {
     case 'travelCard':
       return (
         <View style={styles.container}>
